Saving Reworked
It's becoming clear that my previous method of handling savegames is too tedious for the developer.  You could not jump to partway into the game without losing access to any changes made in the CSVs.  So, I'm trying out a method that involves saving only "updates" and then when loading all updates are applied.  At the moment this only works from the base game.  If this ends up being an improvement, then I will merge the branch to main.