Fixed actionsDisabled! And other bug fixes
Fixed the game-breaking bug! (Actions were being disabled properly, but were for no apparent reason being re-enabled on menu display). And the order of game objects being drawn... then the order of calculating hover object.  Put a delay after displaying the title art.  Added witch art stolen from the title screen. Put a gdignore in Saves so it doesn't try to load as translations.  Put a check in the input handler so it stops reporting all the null inputs.